Resolve Conflicts:
	server/src-lib/Hasura/RQL/DML/Insert.hs
	server/tests-py/queries/graphql_mutation/insert/permissions/author_user_role_insert_check_perm_success.yaml
	server/tests-py/queries/graphql_mutation/insert/permissions/company_user_role_on_conflict.yaml

+++ console/src/components/Services/EventTrigger/Settings/Settings.js
@@ -20,8 +20,12 @@ class Settings extends Component {
     )[0];
     triggerSchema = triggerSchema ? triggerSchema : {};
 
-    const handleDeleteTrigger = () => {
-      dispatch(deleteTrigger(triggerName));
+    const handleDeleteTrigger = e => {
+      e.preventDefault();
+      const isOk = confirm('Are you sure?');
+      if (isOk) {
+        dispatch(deleteTrigger(triggerName));
+      }
     };
 
     return (
